+++ server/src/ffmpeg/builder/MediaStream.test.ts
@@ -8,11 +8,43 @@ describe('MediaStream', () => {
       codec: 'h264',
       frameSize: FrameSize.withDimensions(720, 480),
       index: 0,
-      sampleAspectRatio: null,
+      providedSampleAspectRatio: null,
       displayAspectRatio: '1.33',
       pixelFormat: new PixelFormatYuv420P(),
     });
 
-    console.log(stream.squarePixelFrameSize(FrameSize.FHD));
+    expect(stream.squarePixelFrameSize(FrameSize.FHD)).toMatchFrameSize(
+      FrameSize.withDimensions(1621, 1080),
+    );
+  });
+
+  test('derives sampleAspectRatio with decimal DAR', () => {
+    const stream = VideoStream.create({
+      codec: 'h264',
+      frameSize: FrameSize.withDimensions(720, 480),
+      index: 0,
+      providedSampleAspectRatio: '0:0',
+      displayAspectRatio: '1.33',
+      pixelFormat: new PixelFormatYuv420P(),
+    });
+
+    expect(stream.isAnamorphic).toBeTruthy();
+    expect(stream.sampleAspectRatio).toEqual(
+      `${(1.33).toFixed(12)}:${(1.5).toFixed(12)}`,
+    );
+  });
+
+  test('derives sampleAspectRatio with ratio DAR', () => {
+    const stream = VideoStream.create({
+      codec: 'h264',
+      frameSize: FrameSize.withDimensions(720, 480),
+      index: 0,
+      providedSampleAspectRatio: '0:0',
+      displayAspectRatio: '16:9',
+      pixelFormat: new PixelFormatYuv420P(),
+    });
+
+    expect(stream.isAnamorphic).toBeTruthy();
+    expect(stream.sampleAspectRatio).toEqual(`16:${(1.5).toFixed(12)}`);
   });
 });