+++ server/src/ffmpeg/builder/pipeline/hardware/VaapiPipelineBuilder.test.ts
@@ -46,7 +46,7 @@ describe('VaapiPipelineBuilder', () => {
         frameSize: FrameSize.withDimensions(1920, 900),
         index: 0,
         pixelFormat: new PixelFormatYuv420P(),
-        sampleAspectRatio: null,
+        providedSampleAspectRatio: null,
       }),
     );
 
@@ -130,7 +130,7 @@ describe('VaapiPipelineBuilder', () => {
         frameSize: FrameSize.withDimensions(1920, 900),
         index: 0,
         pixelFormat: new PixelFormatYuv420P(),
-        sampleAspectRatio: null,
+        providedSampleAspectRatio: null,
       }),
     );
 
@@ -216,7 +216,7 @@ describe('VaapiPipelineBuilder', () => {
         frameSize: FrameSize.withDimensions(1920, 900),
         index: 0,
         pixelFormat: new PixelFormatYuv420P(),
-        sampleAspectRatio: null,
+        providedSampleAspectRatio: null,
       }),
     );
 
@@ -303,7 +303,7 @@ describe('VaapiPipelineBuilder', () => {
         frameSize: FrameSize.withDimensions(1920, 900),
         index: 0,
         pixelFormat: new PixelFormatYuv420P(),
-        sampleAspectRatio: null,
+        providedSampleAspectRatio: null,
       }),
     );