+++ CHANGELOG.md
@@ -338,6 +338,43 @@
 * allow selecting parent Jellyfin items in list view ([9756e35](https://github.com/chrisbenincasa/tunarr/commit/9756e35008e36b0acb95f726bebd04f6ed706c78))
 * **streaming:** change how QSV is initialized on Windows ([1dbec53](https://github.com/chrisbenincasa/tunarr/commit/1dbec5357bed545869a685b02bfa5440a3649243))
 
+## [0.22.16](https://github.com/chrisbenincasa/tunarr/compare/v0.22.15...v0.22.16) (2025-11-24)
+
+
+### Bug Fixes
+
+* dynamically calculate SAR when necessary ([69c6947](https://github.com/chrisbenincasa/tunarr/commit/69c6947d081193698afb10ae5c8a1b5058c1cfbc))
+* fix build with new sample aspect ratio fields ([80fccd2](https://github.com/chrisbenincasa/tunarr/commit/80fccd257a88158aff7715bead951a80a92c7a83))
+* fix tests after SAR changes ([944039e](https://github.com/chrisbenincasa/tunarr/commit/944039e00e5826c8f5bba4e94662e6baf5b3010a))
+* make duration_ts option on video streams in ffprobe output ([fe34da8](https://github.com/chrisbenincasa/tunarr/commit/fe34da871e5959269d99bab0879f8260e332d51d))
+* **streaming:** properly set decoder thread count to 1 for hwaccel pipelines ([72f128b](https://github.com/chrisbenincasa/tunarr/commit/72f128b39132f469561773f1b5cc554d201df01c))
+* **ui:** fix remaining time calculation and display in guide ([3c5ea7a](https://github.com/chrisbenincasa/tunarr/commit/3c5ea7a5780587d72375913f1f0efd0b48a3721f))
+
+## [0.22.15](https://github.com/chrisbenincasa/tunarr/compare/v0.22.14...v0.22.15) (2025-11-16)
+
+
+### Bug Fixes
+
+* do not pass exact duration to audio pad filter ([289acf1](https://github.com/chrisbenincasa/tunarr/commit/289acf1c009da52396b440db9626f0f31df5a4d2))
+* fix bad form state in channel EPG settings ([ec82018](https://github.com/chrisbenincasa/tunarr/commit/ec820188a3c50e254668818cd61c6385028be104)), closes [#1468](https://github.com/chrisbenincasa/tunarr/issues/1468)
+* more balanced cyclic shuffle implementation ([0ceb5e3](https://github.com/chrisbenincasa/tunarr/commit/0ceb5e383705fc548cecea6574bb78d48494cda7))
+* remove BackfillProgramGroupings fixer ([f6cf17f](https://github.com/chrisbenincasa/tunarr/commit/f6cf17f27000500b5b71a67e99318696c35436c2))
+
+## [0.22.14](https://github.com/chrisbenincasa/tunarr/compare/v0.22.13...v0.22.14) (2025-11-10)
+
+
+### Bug Fixes
+
+* overhaul logs and system logs page ([24b3213](https://github.com/chrisbenincasa/tunarr/commit/24b32130789c0c3666eefa2dc43113db659833ce))
+
+## [0.22.13](https://github.com/chrisbenincasa/tunarr/compare/v0.22.12...v0.22.13) (2025-11-10)
+
+
+### Bug Fixes
+
+* disable time slot editor buttong when there are no slots ([cc4f056](https://github.com/chrisbenincasa/tunarr/commit/cc4f056e559f0e6d13ec6b7fc1403dc4807b9c55))
+* fix sidear ASS subtitle download in Jellyfin ([5784079](https://github.com/chrisbenincasa/tunarr/commit/578407965ce473006abbd65ea1b3c0ca0201e404)), closes [#1457](https://github.com/chrisbenincasa/tunarr/issues/1457)
+
 ## [0.22.12](https://github.com/chrisbenincasa/tunarr/compare/v0.22.11...v0.22.12) (2025-11-06)